Separated out tests for checking git status.
